pyproject.toml
[project]
name = "uoe-agents"
version = "0.1.0"
description = "A tutorial on building AI agents with Pydantic AI."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic-ai-slim[mcp,openai]>=1.67.0",
]
